pyproject.toml
[tool.poetry]
name = "bpy-addon-build"
version = "0.2.0"
description = "A build system to make building and testing Blender addons 10 times easier"
authors = ["StandingPad"]
license = "MIT"
homepage = "https://github.com/StandingPadAnimations/bpy-build"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^23.1.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.4.1"

[tool.poetry.scripts]
bpy-addon-build = "bpy_addon_build:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
